core: consolidate floating point types used for stats
Fixes problems identified with -Wfloat-conversion

+++ rosmon_core/src/launch/node.cpp
@@ -152,7 +152,7 @@ void Node::setMemoryLimit(uint64_t memoryLimitByte)
 	m_memoryLimitByte = memoryLimitByte;
 }
 
-void Node::setCPULimit(float cpuLimit)
+void Node::setCPULimit(double cpuLimit)
 {
 	m_cpuLimit = cpuLimit;
 }
